amptrade-websocket/requirements.txt
NorenRestApi==0.0.32
websockets
python-dotenv
requests
numpy
fastapi
uvicorn
psutil
aiohttp
protobuf>=6.31.1